requirements.txt
langchain-groq
langchain-huggingface
langchain-chroma
langchain-community
langchain-text-splitters
langchain-core
langgraph
sentence-transformers
chromadb
pypdf
python-dotenv
